signature: Start using RwLock for ParsecProviderSignatureContext
With the current scenario, in which ParsecProviderSignatureContext
is used through an Arc, if we want to edit the components of
ParsecProviderSignatureContext we need to have them wrapped inside
a Mutex.

Start using ParsecProviderSignatureContext through an Arc<RwLock
ParsecProviderSignatureContext>> so as to be able to have multiple
references of it and be able to edit the entire structure fields
in a synchronized way.

Signed-off-by: Tomás González <tomasagustin.gonzalezorlando@arm.com>